Change all database Ids to longs. Rework the look of group job actions. Add ability to enable or disable jobs in batches

+++ SSAReplacement.Api/Domain/Executables/ExecutableParameter.cs
@@ -3,7 +3,7 @@
 public class ExecutableParameter
 {
     public long Id { get; set; }
-    public int ExecutableVersionId { get; set; }
+    public long ExecutableVersionId { get; set; }
     public string Name { get; set; } = "";
     public string TypeName { get; set; } = "";
     public string? Description { get; set; }
