+++ SSAReplacement.Api/Domain/JobRuns/JobRun.cs
@@ -2,15 +2,15 @@ namespace SSAReplacement.Api.Domain;
 
 public class JobRun
 {
-    public int Id { get; set; }
-    public int JobId { get; set; }
-    public int ExecutableVersionId { get; set; }
-    public int? ScheduleId { get; set; }
+    public long Id { get; set; }
+    public long JobId { get; set; }
+    public long ExecutableVersionId { get; set; }
+    public long? ScheduleId { get; set; }
     public DateTime StartedAt { get; set; }
     public DateTime? FinishedAt { get; set; }
-    public required string Status { get; set; } // Running, Success, Failed
+    public required string Status { get; set; }
     public int? ExitCode { get; set; }
-    public string? Trigger { get; set; } // Scheduled, Manual
+    public string? Trigger { get; set; }
 
     public Job Job { get; set; } = null!;
     public ExecutableVersion ExecutableVersion { get; set; } = null!;
