+++ SSAReplacement.Api/Domain/Jobs/JobLog.cs
@@ -2,10 +2,10 @@ namespace SSAReplacement.Api.Domain;
 
 public class JobLog
 {
-    public int Id { get; set; }
-    public int JobRunId { get; set; }
+    public long Id { get; set; }
+    public long JobRunId { get; set; }
     public DateTime LogDate { get; set; }
-    public required string LogType { get; set; } // StdOut, StdErr
+    public required string LogType { get; set; }
     public required string Content { get; set; }
 
     public JobRun JobRun { get; set; } = null!;
