+++ SSAReplacement.Api/Domain/Jobs/JobSchedule.cs
@@ -2,9 +2,9 @@ namespace SSAReplacement.Api.Domain;
 
 public class JobSchedule
 {
-    public int Id { get; set; }
-    public int JobId { get; set; }
-    public int ScheduleId { get; set; }
+    public long Id { get; set; }
+    public long JobId { get; set; }
+    public long ScheduleId { get; set; }
 
     public Job Job { get; set; } = null!;
     public Schedule Schedule { get; set; } = null!;
