+++ SSAReplacement.Api/Features/Executables/Domain/ExecutableDto.cs
@@ -2,14 +2,14 @@ using SSAReplacement.Api.Domain;
 
 namespace SSAReplacement.Api.Features.Executables.Domain;
 
-public record ExecutableDto(int Id, string Name, DateTime CreatedAt, int ActiveVersion)
+public record ExecutableDto(long Id, string Name, DateTime CreatedAt, int ActiveVersion)
 {
     public static ExecutableDto From(Executable e) => new(
         e.Id, e.Name, e.CreatedAt,
         e.Versions.FirstOrDefault(v => v.IsActive)?.Version ?? 0);
 }
 
-public record ExecutableDetailDto(int Id, string? Name, DateTime CreatedAt, IReadOnlyList<ExecutableVersionDto> Versions)
+public record ExecutableDetailDto(long Id, string? Name, DateTime CreatedAt, IReadOnlyList<ExecutableVersionDto> Versions)
 {
     public static ExecutableDetailDto From(Executable e) => new(
         e.Id, e.Name, e.CreatedAt,
