+++ SSAReplacement.Api/Features/Executables/Domain/ExecutableParameterDto.cs
@@ -2,7 +2,7 @@ using SSAReplacement.Api.Domain;
 
 namespace SSAReplacement.Api.Features.Executables.Domain;
 
-public record ExecutableParameterDto(long Id, int ExecutableVersionId, string Name, string TypeName, string? Description, bool Required, string? DefaultValue)
+public record ExecutableParameterDto(long Id, long ExecutableVersionId, string Name, string TypeName, string? Description, bool Required, string? DefaultValue)
 {
     public static ExecutableParameterDto From(ExecutableParameter p) => new(
         p.Id,
