+++ SSAReplacement.Api/Features/Schedules/Handlers/GetScheduleById.cs
@@ -6,7 +6,7 @@ namespace SSAReplacement.Api.Features.Schedules.Handlers;
 
 public static class GetScheduleById
 {
-    public static async Task<IResult> Handler(int id, AppDbContext db)
+    public static async Task<IResult> Handler(long id, AppDbContext db)
     {
         var s = await db.Schedules.AsNoTracking().FirstOrDefaultAsync(x => x.Id == id);
         return s is null ? Results.NotFound() : Results.Ok(ScheduleDto.From(s));
