+++ SSAReplacement.Api/Infrastructure/Migrations/20250228000000_InitialSchema.cs
@@ -8,23 +8,23 @@ public sealed class InitialSchema : Migration
     public override void Up()
     {
         Create.Table("Job")
-            .WithColumn("Id").AsInt32().PrimaryKey().Identity()
-            .WithColumn("ExecutableId").AsInt32().NotNullable()
+            .WithColumn("Id").AsInt64().PrimaryKey().Identity()
+            .WithColumn("ExecutableId").AsInt64().NotNullable()
             .WithColumn("Name").AsString(256).NotNullable()
             .WithColumn("IsEnabled").AsBoolean().NotNullable()
             .WithColumn("CreatedAt").AsDateTime2().NotNullable()
             .WithColumn("NotifyEmail").AsString(256).Nullable();
 
         Create.Table("JobVariable")
-            .WithColumn("Id").AsInt32().PrimaryKey().Identity()
-            .WithColumn("JobId").AsInt32().NotNullable()
+            .WithColumn("Id").AsInt64().PrimaryKey().Identity()
+            .WithColumn("JobId").AsInt64().NotNullable()
             .WithColumn("Key").AsString(256).NotNullable()
             .WithColumn("Value").AsString(4096).NotNullable();
 
         Create.Table("JobRun")
-            .WithColumn("Id").AsInt32().PrimaryKey().Identity()
-            .WithColumn("JobId").AsInt32().NotNullable()
-            .WithColumn("ExecutableVersionId").AsInt32().NotNullable()
+            .WithColumn("Id").AsInt64().PrimaryKey().Identity()
+            .WithColumn("JobId").AsInt64().NotNullable()
+            .WithColumn("ExecutableVersionId").AsInt64().NotNullable()
             .WithColumn("ScheduleId").AsInt32().Nullable()
             .WithColumn("StartedAt").AsDateTime2().NotNullable()
             .WithColumn("FinishedAt").AsDateTime2().Nullable()
@@ -33,40 +33,40 @@ public sealed class InitialSchema : Migration
             .WithColumn("Trigger").AsString(32).Nullable();
 
         Create.Table("JobLog")
-            .WithColumn("Id").AsInt32().PrimaryKey().Identity()
-            .WithColumn("JobRunId").AsInt32().NotNullable()
+            .WithColumn("Id").AsInt64().PrimaryKey().Identity()
+            .WithColumn("JobRunId").AsInt64().NotNullable()
             .WithColumn("LogDate").AsDateTime2().NotNullable()
             .WithColumn("LogType").AsString(16).NotNullable()
             .WithColumn("Content").AsString().NotNullable();
 
         Create.Table("JobSchedule")
-            .WithColumn("Id").AsInt32().PrimaryKey().Identity()
-            .WithColumn("JobId").AsInt32().NotNullable()
-            .WithColumn("ScheduleId").AsInt32().NotNullable();
+            .WithColumn("Id").AsInt64().PrimaryKey().Identity()
+            .WithColumn("JobId").AsInt64().NotNullable()
+            .WithColumn("ScheduleId").AsInt64().NotNullable();
 
         Create.Table("Schedule")
-           .WithColumn("Id").AsInt32().PrimaryKey().Identity()
+           .WithColumn("Id").AsInt64().PrimaryKey().Identity()
            .WithColumn("Name").AsString(256).NotNullable()
            .WithColumn("CronExpression").AsString(256).NotNullable()
            .WithColumn("IsEnabled").AsBoolean().NotNullable()
            .WithColumn("CreatedAt").AsDateTime2().NotNullable();
 
         Create.Table("Executable")
-            .WithColumn("Id").AsInt32().PrimaryKey().Identity()
+            .WithColumn("Id").AsInt64().PrimaryKey().Identity()
             .WithColumn("Name").AsString(256).NotNullable()
             .WithColumn("CreatedAt").AsDateTime2().NotNullable();
 
         Create.Table("ExecutableVersion")
-            .WithColumn("Id").AsInt32().PrimaryKey().Identity()
-            .WithColumn("ExecutableId").AsInt32().NotNullable()
+            .WithColumn("Id").AsInt64().PrimaryKey().Identity()
+            .WithColumn("ExecutableId").AsInt64().NotNullable()
             .WithColumn("Version").AsInt32().NotNullable()
             .WithColumn("EntryPointDll").AsString(256).NotNullable()
             .WithColumn("UploadedAt").AsDateTime2().NotNullable()
             .WithColumn("IsActive").AsBoolean().NotNullable();
 
         Create.Table("ExecutableParameter")
-            .WithColumn("Id").AsInt32().PrimaryKey().Identity()
-            .WithColumn("ExecutableVersionId").AsInt32().NotNullable()
+            .WithColumn("Id").AsInt64().PrimaryKey().Identity()
+            .WithColumn("ExecutableVersionId").AsInt64().NotNullable()
             .WithColumn("Name").AsString(256).NotNullable()
             .WithColumn("TypeName").AsString(256).NotNullable()
             .WithColumn("Description").AsString(1024).Nullable()
